requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
av==16.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
coloredlogs==15.0.1
ctranslate2==4.7.1
edge-tts==7.2.7
exceptiongroup==1.3.1
fastapi==0.135.1
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
numba==0.64.0
numpy==2.2.6
onnxruntime==1.23.2
openai-whisper==20250625
packaging==26.0
piper-tts==1.4.1
propcache==0.4.1
protobuf==7.34.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
shellingham==1.5.4
sounddevice==0.5.5
starlette==0.52.1
sympy==1.14.0
tabulate==0.10.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
whisper-ctranslate2==0.5.7
yarl==1.23.0
yt-dlp==2026.3.3
